siwe: fix rpx error handling - throws and handles rpc error - no longer throws uncaught error
